Refactor ReleasesViewModel
